requirements.txt
google-api-python-client==2.192.0
google-auth-oauthlib==1.3.0
transformers==5.3.0
torch==2.10.0
requests==2.32.5
python-dotenv==1.2.2
